+++ src/Features/Editor/Schedule/Schedule.module.scss
@@ -6,25 +6,6 @@
   height: calc(100vh - 12.25rem);
 }
 
-.listContainer {
-  background: var(--cds-layer);
-  width: 31.725rem;
-  flex-shrink: 0;
-  padding: 2rem 1rem;
-  display: flex;
-  flex-direction: column;
-  overflow: hidden;
-  gap: 1rem;
-
-  h2 {
-    font-size: 1.25rem;
-  }
-
-  ul {
-    overflow-y: scroll;
-  }
-}
-
 .calendarContainer {
   flex-grow: 1;
   padding: 1.5rem;
@@ -36,127 +17,5 @@
 }
 
 .panelContainer {
-  background-color: var(--cds-ui-01);
   margin-top: 12.125rem;
-  border: none;
-  height: calc(100vh - 12.125rem);
-  transition: transform $carbon-motion--timing--productive $carbon-motion--easing--productive;
-  :global .slide-pane__content {
-    padding: 0;
-  }
-  h2 {
-    font-size: 1.25rem;
-    margin-bottom: 1rem;
-  }
-}
-
-.listItem {
-  background: var(--cds-ui-background);
-  border-radius: 0.25rem;
-  min-height: 8rem;
-  margin: 1rem 0rem;
-  padding: 1rem;
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-  font-size: 0.9375rem;
-
-  dt {
-    font-size: 0.75rem;
-  }
-  dd {
-    font-size: 0.875rem;
-  }
-}
-
-.listItemTitle {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  max-width: 90%;
-
-  h3 {
-    font-size: 1rem;
-    font-weight: bold;
-    @include text-overflow-ellipsis();
-  }
-}
-
-.listItemDescription {
-  @include multiline-text-overflow-ellipsis();
-  max-width: 90%;
-}
-
-.listItemLabels {
-  display: flex;
-  max-width: 90%;
-  span:first-child {
-    margin-left: 0rem;
-  }
-}
-
-.statusCircle {
-  height: 1rem;
-  width: 1rem;
-  &[data-status="active"] {
-    fill: var(--cds-support-success);
-  }
-  &[data-status="inactive"] {
-    fill: var(--cds-support-success);
-  }
-  &[data-status="deleted"] {
-    fill: var(--cds-support-error);
-  }
-
-  &[data-status="trigger_disabled"] {
-    fill: var(--cds-text-03);
-  }
-}
-
-.detailsSection {
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-
-  h2 {
-    display: flex;
-    font-weight: bold;
-    margin-bottom: 0rem;
-  }
-
-  dl {
-    margin-top: 1rem;
-  }
-
-  dt {
-    font-size: 0.875rem;
-  }
-
-  section {
-    margin: 1rem;
-    padding-bottom: 1.5rem;
-    border-bottom: 1px var(--cds-ui-04) solid;
-  }
-
-  footer {
-    margin-top: auto;
-    width: 100%;
-    button {
-      border-radius: 0;
-      width: 50% !important;
-    }
-  }
-}
-
-.detailsTitle {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  width: 100%;
-
-  h2 {
-    max-width: 80%;
-    @include text-overflow-ellipsis();
-  }
 }
