+++ src/Features/Schedule/Schedule.module.scss
@@ -87,84 +87,7 @@
   z-index: 0;
 }
 
-.listItem {
-  background: var(--cds-ui-background);
-  border-radius: 0.25rem;
-  min-height: 8rem;
-  margin: 1rem 0rem;
-  padding: 1rem;
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-  font-size: 0.9375rem;
-
-  dt {
-    font-size: 0.75rem;
-  }
-  dd {
-    font-size: 0.875rem;
-  }
-}
-
-.listItemTitle {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  max-width: 90%;
-
-  h3 {
-    font-size: 1rem;
-    font-weight: bold;
-    @include text-overflow-ellipsis();
-  }
-}
-
-.listItemDescription {
-  @include multiline-text-overflow-ellipsis();
-  max-width: 90%;
-}
-
-.listItemLabels {
-  display: flex;
-  max-width: 90%;
-}
-
-.statusCircle {
-  height: 1rem;
-  width: 1rem;
-  &[data-status="active"] {
-    fill: var(--cds-support-success);
-  }
-  &[data-status="inactive"] {
-    fill: var(--cds-support-success);
-  }
-  &[data-status="deleted"] {
-    fill: var(--cds-support-error);
-  }
-
-  &[data-status="trigger_disabled"] {
-    fill: var(--cds-text-03);
-  }
-}
-
-.detailsSection {
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-
-  section {
-    margin: 1rem;
-    padding-bottom: 1.5rem;
-    border-bottom: 1px var(--cds-ui-04) solid;
-  }
-
-  footer {
-    margin-top: auto;
-    width: 100%;
-    button {
-      border-radius: 0;
-      width: 50% !important;
-    }
-  }
+.panelContainer {
+  margin-top: 9.875rem;
+  height: calc(100vh - 9.875rem);
 }
